OpenStack Heat cluster autoscaler
Implementation of an OpenStack cloudprovider using Heat for the
cluster autoscaler. Here, clusters are necessarily thought of as Heat
ResourceGroups or AutoScalingGroups. The scaling is done through
parameter updates of the Heat stacks.

The commit contains no vendored dependencies.

This is based on work from Atanas Mirchev <atanas.mirchev@est.fujitsu.com>
https://github.com/kubernetes/contrib/pull/1980